+++ packages/libs/react-ui/src/components/Form/InputWrapper/InputWrapper.stories.tsx
@@ -1,7 +1,6 @@
+import type { IInputWrapperProps } from '@components/Form';
+import { Input, InputWrapper } from '@components/Form';
 import type { SystemIcon } from '@components/Icon';
-import { Input } from '@components/Input';
-import type { IInputWrapperProps } from '@components/InputWrapper';
-import { InputWrapper } from '@components/InputWrapper';
 import type { Meta, StoryObj } from '@storybook/react';
 import { vars } from '@theme/vars.css';
 import React from 'react';
