+++ packages/libs/react-ui/src/components/Form/SelectField/SelectField.stories.tsx
@@ -1,25 +1,21 @@
+import type { ISelectFieldProps } from '@components/Form';
+import { SelectField } from '@components/Form';
+import { statusVariant } from '@components/Form/InputWrapper/InputWrapper.css';
 import { SystemIcon } from '@components/Icon';
-import { statusVariant } from '@components/InputWrapper/InputWrapper.css';
-import type { ITextFieldProps } from '@components/TextField';
-import { TextField } from '@components/TextField';
 import type { Meta, StoryObj } from '@storybook/react';
-import { vars } from '@theme/vars.css';
 import React from 'react';
 
 type StoryProps = {
-  helperText: string;
-  leadingText: string;
-  leftIcon: keyof typeof SystemIcon;
-  rightIcon: keyof typeof SystemIcon;
-} & ITextFieldProps;
+  icon: keyof typeof SystemIcon;
+} & ISelectFieldProps;
 
 const meta: Meta<StoryProps> = {
-  title: 'Form/TextField',
+  title: 'Components/SelectField',
   parameters: {
     docs: {
       description: {
         component:
-          'TextField is the composition of Input and InputWrapper to provide an input with a label, helper text, and other peripheral information.',
+          'SelectField is the composition of the Select and InputWrapper components to provide a select with a label, helper text, and other peripheral information.',
       },
     },
   },
@@ -49,23 +45,6 @@ const meta: Meta<StoryProps> = {
         type: 'text',
       },
     },
-    leadingText: {
-      description: "Leading text that is rendered inside the input's border.",
-      control: {
-        type: 'text',
-      },
-    },
-    leadingTextWidth: {
-      description:
-        'Width of the leading text. Defaults to the size of the text itself.',
-      control: {
-        type: 'select',
-      },
-      options: [
-        undefined,
-        ...Object.keys(vars.sizes).map((key) => key as keyof typeof vars.sizes),
-      ],
-    },
     status: {
       options: [
         undefined,
@@ -87,17 +66,8 @@ const meta: Meta<StoryProps> = {
         defaultValue: { summary: 'false' },
       },
     },
-    leftIcon: {
-      description:
-        'Icon rendered inside the input to the left of the input text.',
-      options: Object.keys(SystemIcon) as (keyof typeof SystemIcon)[],
-      control: {
-        type: 'select',
-      },
-    },
-    rightIcon: {
-      description:
-        'Icon rendered inside the input to the right of the input text.',
+    icon: {
+      description: 'Icon rendered inside the select to the left of the text.',
       options: Object.keys(SystemIcon) as (keyof typeof SystemIcon)[],
       control: {
         type: 'select',
@@ -108,14 +78,8 @@ const meta: Meta<StoryProps> = {
 
 type Story = StoryObj<StoryProps>;
 
-/*
- *👇 Render functions are a framework specific feature to allow you control on how the component renders.
- * See https://storybook.js.org/docs/7.0/react/api/csf
- * to learn how to use render functions.
- */
-
 export const Group: Story = {
-  name: 'Text Field',
+  name: 'Select Field',
   args: {
     tag: 'tag',
     helperText: 'This is helper text',
@@ -123,40 +87,27 @@ export const Group: Story = {
     label: 'Label',
     disabled: false,
     status: undefined,
-    leftIcon: 'Account',
-    rightIcon: undefined,
-    leadingText: 'Leading',
-    leadingTextWidth: undefined,
+    icon: 'Account',
   },
-  render: ({
-    leadingText,
-    leftIcon,
-    rightIcon,
-    disabled,
-    status,
-    tag,
-    helperText,
-    info,
-    label,
-    leadingTextWidth,
-  }) => {
+  render: ({ icon, disabled, status, tag, helperText, info, label }) => {
     return (
-      <TextField
+      <SelectField
         tag={tag}
         info={info}
         label={label}
         status={status}
         disabled={disabled}
         helperText={helperText}
-        leadingTextWidth={leadingTextWidth}
-        inputProps={{
+        selectProps={{
+          ariaLabel: 'Select Story',
           id: 'inputStory',
-          leadingText,
-          leftIcon,
-          rightIcon,
+          icon,
           placeholder: 'This is a placeholder',
         }}
-      />
+      >
+        <option value="1">Option 1</option>
+        <option value="2">Option 2</option>
+      </SelectField>
     );
   },
 };
