+++ packages/libs/react-ui/src/components/Form/TextField/TextField.stories.tsx
@@ -1,21 +1,26 @@
+import type { ITextFieldProps } from '@components/Form';
+import { TextField } from '@components/Form';
+// what do we want this to look like
+import { statusVariant } from '@components/Form/InputWrapper/InputWrapper.css';
 import { SystemIcon } from '@components/Icon';
-import { statusVariant } from '@components/InputWrapper/InputWrapper.css';
-import type { ISelectFieldProps } from '@components/SelectField';
-import { SelectField } from '@components/SelectField';
 import type { Meta, StoryObj } from '@storybook/react';
+import { vars } from '@theme/vars.css';
 import React from 'react';
 
 type StoryProps = {
-  icon: keyof typeof SystemIcon;
-} & ISelectFieldProps;
+  helperText: string;
+  leadingText: string;
+  leftIcon: keyof typeof SystemIcon;
+  rightIcon: keyof typeof SystemIcon;
+} & ITextFieldProps;
 
 const meta: Meta<StoryProps> = {
-  title: 'Components/SelectField',
+  title: 'Form/TextField',
   parameters: {
     docs: {
       description: {
         component:
-          'SelectField is the composition of the Select and InputWrapper components to provide a select with a label, helper text, and other peripheral information.',
+          'TextField is the composition of Input and InputWrapper to provide an input with a label, helper text, and other peripheral information.',
       },
     },
   },
@@ -45,6 +50,23 @@ const meta: Meta<StoryProps> = {
         type: 'text',
       },
     },
+    leadingText: {
+      description: "Leading text that is rendered inside the input's border.",
+      control: {
+        type: 'text',
+      },
+    },
+    leadingTextWidth: {
+      description:
+        'Width of the leading text. Defaults to the size of the text itself.',
+      control: {
+        type: 'select',
+      },
+      options: [
+        undefined,
+        ...Object.keys(vars.sizes).map((key) => key as keyof typeof vars.sizes),
+      ],
+    },
     status: {
       options: [
         undefined,
@@ -66,8 +88,17 @@ const meta: Meta<StoryProps> = {
         defaultValue: { summary: 'false' },
       },
     },
-    icon: {
-      description: 'Icon rendered inside the select to the left of the text.',
+    leftIcon: {
+      description:
+        'Icon rendered inside the input to the left of the input text.',
+      options: Object.keys(SystemIcon) as (keyof typeof SystemIcon)[],
+      control: {
+        type: 'select',
+      },
+    },
+    rightIcon: {
+      description:
+        'Icon rendered inside the input to the right of the input text.',
       options: Object.keys(SystemIcon) as (keyof typeof SystemIcon)[],
       control: {
         type: 'select',
@@ -78,8 +109,14 @@ const meta: Meta<StoryProps> = {
 
 type Story = StoryObj<StoryProps>;
 
+/*
+ *👇 Render functions are a framework specific feature to allow you control on how the component renders.
+ * See https://storybook.js.org/docs/7.0/react/api/csf
+ * to learn how to use render functions.
+ */
+
 export const Group: Story = {
-  name: 'Select Field',
+  name: 'Text Field',
   args: {
     tag: 'tag',
     helperText: 'This is helper text',
@@ -87,27 +124,40 @@ export const Group: Story = {
     label: 'Label',
     disabled: false,
     status: undefined,
-    icon: 'Account',
+    leftIcon: 'Account',
+    rightIcon: undefined,
+    leadingText: 'Leading',
+    leadingTextWidth: undefined,
   },
-  render: ({ icon, disabled, status, tag, helperText, info, label }) => {
+  render: ({
+    leadingText,
+    leftIcon,
+    rightIcon,
+    disabled,
+    status,
+    tag,
+    helperText,
+    info,
+    label,
+    leadingTextWidth,
+  }) => {
     return (
-      <SelectField
+      <TextField
         tag={tag}
         info={info}
         label={label}
         status={status}
         disabled={disabled}
         helperText={helperText}
-        selectProps={{
-          ariaLabel: 'Select Story',
+        leadingTextWidth={leadingTextWidth}
+        inputProps={{
           id: 'inputStory',
-          icon,
+          leadingText,
+          leftIcon,
+          rightIcon,
           placeholder: 'This is a placeholder',
         }}
-      >
-        <option value="1">Option 1</option>
-        <option value="2">Option 2</option>
-      </SelectField>
+      />
     );
   },
 };
