+++ packages/libs/react-ui/src/components/index.ts
@@ -8,11 +8,19 @@ export type { IBreadcrumbItemProps, IBreadcrumbsProps } from './Breadcrumbs';
 export type { IButtonProps } from './Button';
 export type { ICardProps } from './Card';
 export type { IContentHeaderProps } from './ContentHeader';
+export type {
+  IInputProps,
+  IInputWrapperProps,
+  ISelectFieldProps,
+  ISelectProps,
+  ITextAreaFieldProps,
+  ITextFieldProps,
+  ITextareaProps,
+  InputWrapperStatus,
+} from './Form';
 export type { IGridContainerProps, IGridItemProps } from './Grid';
 export type { IIconProps } from './Icon';
 export type { IIconButtonProps } from './IconButton';
-export type { IInputProps } from './Input/Input';
-export type { IInputWrapperProps, InputWrapperStatus } from './InputWrapper';
 export type { ILinkProps } from './Link';
 export type { IMaskedValueProps } from './MaskedValue/MaskedValue';
 export type { IModalProps } from './Modal';
@@ -41,8 +49,6 @@ export type {
   IProfileSummaryRootProps,
 } from './ProfileSummary';
 export type { IProgressBarProps } from './ProgressBar';
-export type { ISelectProps } from './Select';
-export type { ISelectFieldProps } from './SelectField';
 export type { IStackProps } from './Stack';
 export type {
   ITBodyProps,
@@ -54,9 +60,6 @@ export type {
 } from './Table';
 export type { ITabContentProps, ITabProps, ITabsContainerProps } from './Tabs';
 export type { ITagProps } from './Tag';
-export type { ITextareaProps } from './TextArea/TextArea';
-export type { ITextAreaFieldProps } from './TextAreaField/TextAreaField';
-export type { ITextFieldProps } from './TextField/TextField';
 export type { ITooltipProps } from './Tooltip';
 export type { ILabelValue, ITrackerCardProps } from './TrackerCard/TrackerCard';
 export type { ITreeProps } from './Tree';
@@ -74,11 +77,18 @@ export { Button } from './Button';
 export { Card } from './Card';
 export { ContentHeader } from './ContentHeader';
 export { Divider } from './Divider/Divider';
+export {
+  Input,
+  InputWrapper,
+  Select,
+  SelectField,
+  TextAreaField,
+  TextField,
+  Textarea,
+} from './Form';
 export { Grid } from './Grid';
 export { ProductIcon, SystemIcon } from './Icon';
 export { IconButton } from './IconButton';
-export { Input } from './Input/Input';
-export { InputWrapper } from './InputWrapper';
 export { Link } from './Link';
 export { MaskedValue } from './MaskedValue/MaskedValue';
 export { Modal, ModalProvider, useModal } from './Modal';
@@ -88,15 +98,10 @@ export { Notification } from './Notification';
 export { Pagination } from './Pagination';
 export { ProfileSummary } from './ProfileSummary';
 export { ProgressBar } from './ProgressBar';
-export { Select } from './Select/Select';
-export { SelectField } from './SelectField/SelectField';
 export { Stack } from './Stack';
 export { Table } from './Table';
 export { Tabs } from './Tabs';
 export { Tag } from './Tag';
-export { Textarea } from './TextArea/TextArea';
-export { TextAreaField } from './TextAreaField/TextAreaField';
-export { TextField } from './TextField/TextField';
 export { Tooltip } from './Tooltip';
 export { TrackerCard } from './TrackerCard/TrackerCard';
 export { Tree } from './Tree';
